--- a/p2p/src/fsm/syncmgr.rs
+++ b/p2p/src/fsm/syncmgr.rs
@@ -570,9 +570,9 @@ impl<U: Wakeup + Disconnect + Wire<Event>, C: Clock> SyncManager<U, C> {
             // TODO: This will be removed.
             Error::BlockImportAborted(_, _, _) => Ok(()),
 
-            // This shouldn't happen here.
+            // These shouldn't happen here.
             // TODO: Perhaps there's a better way to have this error not show up here.
-            Error::Interrupted => Ok(()),
+            Error::Interrupted | Error::GenesisMismatch => Ok(()),
         }
     }