See commits.

On my laptop, mainnet headers now load in 1.2s in debug mode and 100ms in release mode. 
Testnet headers load in 3.8s in debug mode and 300ms in release mode.